Separate Array new Api tests to target netstandard1.7
