+++ src/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -89,6 +89,7 @@
     <Compile Include="$(CommonTestPath)\System\RandomDataGenerator.cs">
       <Link>Common\System\RandomDataGenerator.cs</Link>
     </Compile>
+    <Compile Condition="'$(TargetGroup)'=='netstandard1.7'" Include="Array.1.7.Tests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
